requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
qdrant-client==1.7.0
groq==1.0.0
python-dotenv==1.0.0
